newai/App/requirements.txt
# Core dependencies
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pymongo==4.6.0

# Data handling
pandas==2.0.3
numpy==1.24.3

# Visualizations
plotly==5.15.0
altair==5.2.0

# PDF processing
PyPDF2==3.0.1

# Utilities
requests==2.31.0
python-dateutil==2.8.2
typing-extensions==4.7.1
packaging==23.1
protobuf==4.23.4

# Image processing
Pillow==10.0.0

# Legacy support (if needed)
future==0.18.3